refactor: (#7) mockData 중 notionList => notion 수정
- 해당 파일 내에 노션 아이템 정보만 존재
