+++ mocks/mockData/notion.ts
@@ -17,7 +17,7 @@ export const mockNotionBananaMilk: Notion = {
   id: 2,
   name: '바나나맛 우유',
   description:
-    '바나나맛 우유(영어: Banana Flavored Milk)는 대한민국의 유제품 제조 회사 빙그레에서 1974년 6월에 출시한 바나나 가공유다. 1970년대 정부의 낙농업 육성을 위한 우유 소비 장려 정책에 힘입어 개발되었다. 당시 고급과일의 대명사였던 바나나가 어린이들이 가장 먹고 싶어하는 과일이라는 점에 착안한 제품으로, 항아리 모양의 용기 디자인은 고향을 떠올리게 하기 위한 발상이었다. 당시에는 쥐기 힘들고 보관이 불편하다는 반대 의견도 있었다. 독특한 용기 모양으로 단지우유 라는 애칭으로 불리며 하루 평균 80만개 이상, 연 2억 5만개 이상 판매되고 있다. 1998년 300억원, 2001년 600억원대의 매출을 기록한 데에 이어[4] 가공유 제품으로는 사상 최초로 단일 브랜드로 연매출 1000억원대 기록을 달성했고, 2011년에는 연 1500억원의 매출을 기록했다.[5] 가공유 시장에서 시장점유율 80%를 차지하고 있다. 2004년부터 미국, 캐나다, 중국 등 10여개 국가에 수출되고 있다.[6][7] 국내 유제품 최초로 일본 시장에 진출해 편의점 로손의 8000개 점포에 입점했다. 2006년 건강을 생각하는 젊은층을 타깃으로 지방함량을 1.5% 낮추고 당지수가 낮은 결정과당을 사용한 바나나맛 라이트가 출시된 데에 이어, 2012년 호두, 아몬드 등 견과류의 고소한 맛을 더한 "바나나맛 우유&토피넛"이 출시되었다. 바나나맛 우유&토피넛은 기존 제품과 동일한 용기 디자인에 브랜드 로고의 색깔이 다르다.2018년, "세상에 없던 우유" 시리즈 중 오디맛 우유를 출시했다.',
+    '바나나맛 우유(영어: Banana Flavored Milk)는 대한민국의 유제품 제조 회사 빙그레에서 1974년 6월에 출시한 바나나 가공유다. 1970년대 정부의 낙농업 육성을 위한 우유 소비 장려 정책에 힘입어 개발되었다. \n당시 고급과일의 대명사였던 바나나가 어린이들이 가장 먹고 싶어하는 과일이라는 점에 착안한 제품으로, 항아리 모양의 용기 디자인은 고향을 떠올리게 하기 위한 발상이었다. 당시에는 쥐기 힘들고 보관이 불편하다는 반대 의견도 있었다. \n독특한 용기 모양으로 단지우유 라는 애칭으로 불리며 하루 평균 80만개 이상, 연 2억 5만개 이상 판매되고 있다. 1998년 300억원, 2001년 600억원대의 매출을 기록한 데에 이어 가공유 제품으로는 사상 최초로 단일 브랜드로 연매출 1000억원대 기록을 달성했고, 2011년에는 연 1500억원의 매출을 기록했다. 가공유 시장에서 시장점유율 80%를 차지하고 있다. 2004년부터 미국, 캐나다, 중국 등 10여개 국가에 수출되고 있다. 국내 유제품 최초로 일본 시장에 진출해 편의점 로손의 8000개 점포에 입점했다. \n2006년 건강을 생각하는 젊은층을 타깃으로 지방함량을 1.5% 낮추고 당지수가 낮은 결정과당을 사용한 바나나맛 라이트가 출시된 데에 이어, 2012년 호두, 아몬드 등 견과류의 고소한 맛을 더한 "바나나맛 우유&토피넛"이 출시되었다. 바나나맛 우유&토피넛은 기존 제품과 동일한 용기 디자인에 브랜드 로고의 색깔이 다르다.2018년, "세상에 없던 우유" 시리즈 중 오디맛 우유를 출시했다.',
   relatedNotionList: [
     {
       id: 4,
